Pull out all the tests I wrote for the source map improvement.
Some of them don't actually pass on the pseudo-transformer codepath, so I've commented them out with TODOs to enable them when the relevant transforms have a proper transformer for them.
